+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v1.0.2
+## 11/15/2021
+    1. [](#bugfix)
+    * Fixed an issue with $this->locator->findResource returns false, replaced with $this->locator->isRequested instead
+    * Split up the manifest.json path to get build folder.
+    2. Add vendor folder
+    * Compare to https://github.com/getgrav/grav-plugin-admin, the project now need vendor folder to be commited.
 # v1.0.0
 ## 27-04-2020